Adjust rw Injection
- Needed to adjust friction augmentation based on sign of reaction wheel
speed.
- don't set commanded torque to 0 when rw speed is zero, b/c then it
will always remain zero, regardless of the size of the commanded torque.
Might be a better way to handle this: If we are both below delta_omega
and the commanded torque is sufficiently small, then we command to 0,
otherwise we command at a percentage of the commanded torque (or
something similar).