Don't inject Comments timed injection funcs. THis current state acts nominally.
Also, if we uncomment the fs injetion, we will see a fs fault, and fm
will detect and recover from this fault.

We cannot detect a rw fault because the rw speed data is too noisy to
get an accurate angular acceleration to test against the expected
values. We may find another way to do this, but for now, the timed
injection also intiates the recovery sequence after 20s.